archive: store exact cents — rounding was manufacturing false candidates
Kalshi quotes sub-cent. close_dollars carries four decimal places, so 0.9330 is a
real ask of 93.30c. candle_price rounded it with int(round(x*100)), turning 93.30c
into 93c — which sits inside the live [90,93] band when the true price does not.

The archive is the denominator of every capture figure this repo produces, so the
error propagated into all of them: entries that were never executable were counted as
signals the bot failed to take. An audit on 2026-08-22 probed 180 markets and found
rounding changed 18.4% of selected identities.

candle_price now returns Decimal, not int and not float, so 93.30 stays 93.30 rather
than printing as 93.30000000001. backtest.load parses ask and priors through float().
Days archived before today hold integer cents and parse identically through float(),
so old and new files mix without a migration.

Verified no behaviour drift from the parse change: Aug 18-20 at bet=50 returns
481tr / 92.93%WR / +146, matching the figure measured before this change exactly. The
same window at the live bet now reads +73, which is the $25 sizing, not a regression.

Also added a guard the write path was missing. Kalshi drops settled markets at ~67
days, so a --force re-archive of an old day can legitimately return far less than the
file already holds, and overwriting would destroy data that cannot be re-fetched at
any price. archive_day now refuses to replace an existing archive with one under 95%
of its row count. --force is for upgrading precision, not for truncating history.

Historical days keep their rounded prices until re-archived; that is recoverable
inside Kalshi's retention window and permanently lost outside it.

+++ scripts/backtest.py
@@ -98,7 +98,10 @@ def load(since=None, until=None):
     files = sorted(glob.glob(os.path.join(DATA, "*.csv.gz")))
     if not files:
         sys.exit(f"no data in {DATA} — run scripts/archive_candles.py --backfill N")
-    ip = lambda v: int(v) if v not in ("", "None") else -1
+    # Prices are EXACT cents and may carry decimals (Kalshi quotes sub-cent; a real
+    # ask of 93.30c is not 93c). Days archived before 2026-08-22 hold integer cents,
+    # which parse identically through float(), so old and new files mix safely.
+    ip = lambda v: float(v) if v not in ("", "None") else -1.0
     rows, seen = [], set()
     for path in files:
         day = os.path.basename(path)[:10]
@@ -111,7 +114,7 @@ def load(since=None, until=None):
                     continue
                 seen.add(k)
                 rows.append((r["series"], r["ticker"], int(r["close_ts"]), r["side"],
-                             int(r["ask"]), float(r["secs_left"]), r["won"] == "True",
+                             float(r["ask"]), float(r["secs_left"]), r["won"] == "True",
                              ip(r["prior_1"]), ip(r["prior_2"]), ip(r["prior_3"])))
     if not rows:
         sys.exit("no rows in the requested window")
